refactor: split PR triage into main + GitHub helper modules
Extract GitHub API helpers, data fetching, and utility functions into
pr-triage-github.mjs to keep each file under 400 lines. Update workflow
sparse-checkout to include the new helper module.

- pr-triage.mjs: 393 lines (triage logic, LLM call, labels, main)
- pr-triage-github.mjs: 135 lines (GitHub API, data collection, utilities)
